requirements.txt
# requirements.txt

certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
mega.py==1.0.8
pathlib==1.0.1
psutil==6.1.1
pycryptodome==3.21.0
python-gnupg==0.5.4
requests==2.32.3
six==1.17.0
tenacity==5.1.5
urllib3==2.3.0
